% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getnns.R
\name{getnns}
\alias{getnns}
\title{Nearest neighbor of individuals based on a dissimilarity matrix}
\usage{
getnns(diss, flag)
}
\arguments{
\item{diss}{a dissimilarity matrix}

\item{flag}{a vector of size \code{n} which indicates if we want to compute nearest neighbor of individual i (flag[i]=1) or not (flag[i]=0)}
}
\value{
\item{nn}{a vector of size n which indicates who is the nearest neighbor}

\item{nndiss}{value of the dissimilarity between individuals an their nearest neighbors}
}
\description{
Get nearest neighbor of individuals and the dissimilarity between individuals and his nearest neighbor
}
\keyword{internal}

