% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hclustgeo.uniq.R
\name{hclustgeo.uniq}
\alias{hclustgeo.uniq}
\title{Hierarchical clustering based on geographical distance and on distance between individuals mesured on numerical variables.}
\usage{
hclustgeo.uniq(data, D.geo, alpha)
}
\arguments{
\item{data}{a data frame with \code{n} rows and \code{p} colums containing measures of individuals on variables}

\item{D.geo}{a matrix of size \code{n} by \code{n} containing geographical distances between individuals}

\item{alpha}{a real between 0 and 1 to fit the global criterion to optimize}
}
\value{
{clust}{an object of class hclustgeo.uniq containing the result of hclustgeo}
}
\description{
This version optimize a global criterion of quality of classes H= alpha I_intra_var + (1-alpha) I_intra_geo
}
\keyword{internal}

