% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.hclustgeo.uniq.R
\name{plot.hclustgeo.uniq}
\alias{plot.hclustgeo.uniq}
\title{Differents plots available with the method \code{hclustgeo.uniq}.}
\usage{
\method{plot}{hclustgeo.uniq}(x, choice = "maps", K.range = c(3:5),
  path.shp = NULL, name.ind.shp = NULL, save.pdf = FALSE,
  path.pdf = NULL, ...)
}
\arguments{
\item{x}{an object of class 'hclustgeo.uniq' obtained with \code{hclustgeo.uniq}.}

\item{choice}{a character string: "dendro" if you want to plot the dendrogram; "maps" if you want to plot
the partitions (in K.range clusters) on a map.}

\item{K.range}{a vector containing the different partitions (each K.range correspond to a number of cluster)
you want to represent on a map.}

\item{path.shp}{the path where are the shp and dbf files to plot the map of the typologie.}

\item{name.ind.shp}{the identifier of individuals in the dbf file.}

\item{save.pdf}{a boolean indicating if you want to save the maps obtained in pdf files.}

\item{path.pdf}{if save.pdf=TRUE, the path of the folder where you want to save pdf files.}

\item{...}{other arguments to be passed}
}
\value{
{} {different results and map about the result of the clustering.}
}
\description{
This function provide some graphics and maps obtained with the method \code{hclustgeo.uniq}
}
\keyword{internal}

