% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.hclustgeo.R
\name{print.hclustgeo}
\alias{print.hclustgeo}
\title{Print a 'hclustgeo' object}
\usage{
\method{print}{hclustgeo}(x, ...)
}
\arguments{
\item{x}{an object of class  \code{hclustgeo} generated by the function \code{hclustgeo}}

\item{\dots}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class \code{hclustgeo}.
}

