% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.hclustgeo.uniq.R
\name{print.hclustgeo.uniq}
\alias{print.hclustgeo.uniq}
\title{Print a 'hclustgeo.uniq' object}
\usage{
\method{print}{hclustgeo.uniq}(x, ...)
}
\arguments{
\item{x}{an object of class  \code{hclustgeo.uniq} generated by the function \code{hclustgeo.uniq}}

\item{\dots}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class \code{hclustgeo.uniq}.
}
\keyword{internal}

