% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.summary.hclustgeo.R
\name{print.summary.hclustgeo}
\alias{print.summary.hclustgeo}
\title{Print a 'summary.hclustgeo' object}
\usage{
\method{print}{summary.hclustgeo}(x, ...)
}
\arguments{
\item{x}{an object of class  \code{summary.hclustgeo} generated by the function \code{summary.hclustgeo}}

\item{\dots}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
This is a method for the function print for objects of the class \code{summary.hclustgeo}.
}
\keyword{internal}

