% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.hclustgeo.R
\name{summary.hclustgeo}
\alias{summary.hclustgeo}
\title{Summary of a list of partition obtained with the method \code{hclustgeo}.}
\usage{
\method{summary}{hclustgeo}(object, K.range = c(3, 4, 5), data.desc = NULL,
  ...)
}
\arguments{
\item{object}{an object of class 'hclustgeo' obtained with \code{hclustgeo}.}

\item{K.range}{the maximum number of classes we want the results.}

\item{data.desc}{the database used to describe the different class of the partition. If NULL the database
used for the clustering will be used .}

\item{...}{other arguments}
}
\value{
{} {different results about the result of the clustering.}
}
\description{
This function provide some results of several clustering obtained with \code{hclustgeo}
}

