% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.hclustgeo.uniq.R
\name{summary.hclustgeo.uniq}
\alias{summary.hclustgeo.uniq}
\title{Summary of a partition obtained with the method \code{hclustgeo.uniq}.}
\usage{
\method{summary}{hclustgeo.uniq}(object, K.range = c(3:5), data.desc = NULL,
  ...)
}
\arguments{
\item{object}{an object of class 'hclustgeo.uniq' obtained with \code{hclustgeo.uniq}.}

\item{K.range}{the maximum number of classes we want the results.}

\item{data.desc}{the database used to describe the different class of the partition. If NULL the database
used for the clustering will be used .}

\item{...}{other arguments}
}
\value{
{} {different results about the result of the clustering.}
}
\description{
This function provide some results of a clustering obtained with \code{hclustgeo.uniq}
}
\keyword{internal}

