% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withindiss.R
\name{withindiss}
\alias{withindiss}
\title{Dissimilarity based pseudo within-cluster inertia of a partition}
\usage{
withindiss(D, part, wt = NULL)
}
\arguments{
\item{D}{an object of class "dist" with the dissimilarities between the n observations. 
The function \code{\link{as.dist}} can be used to transform an object of class matrix to object of class "dist".}

\item{part}{a vector with group membership.}

\item{wt}{vector with the weights of the observations}
}
\description{
This function performs the pseudo within-cluster inertia of a partition from a dissimilarity matrix.
}
\keyword{internal}
